refactor(apps-proxy): AJDA-3219 move the reasoning out of the comments
The package carried 376 comment lines against 788 of code, 47%, with
event.go at 266% and sessions.go at 613 lines against the 500-line
guidance in CLAUDE.md. Volume was not the real problem, though:
docs/apps-proxy/sessions.md argues the same points at length, so there
were two copies of every argument, and the first change to either would
have left the other wrong.

Rationale, tradeoffs and cross-system mechanism now live in the doc, and
the comments keep what a reader would otherwise break — shortened, not
deleted: that session_start is gated on the cookie and never on the cache
(by far the most important one, because someone will try to "optimise"
it), buildEvent's lock requirement, the identity-flush critical section,
never closing the writer queue, the fixed timestamp precision, the
keyPrefix collision with the oauth2-proxy key space, SameSite=Lax, that
deadlines never move backwards, and that lastSeen moves only on data
frames.

Down to 280 against 795, 35%.

sessions.go is split rather than shaved further: emit.go takes the half
that records activity and turns it into events, sessions.go keeps the
half that recognises a session and maintains its cookie and cache. 360
and 210 lines, both comfortably under the guidance, and the seam is a
real one.

event.go stays high by ratio at 145% — it is a 16-field struct whose
fields are the target table's columns, and a one-line note per column is
the point of the file rather than an excess.

Fixed in passing: the deadline docblock had been orphaned onto retention
when that helper was inserted above it.

Raised in review by @pepamartinec, whose list of what to keep this
follows.
